Fix a couple more 64-bit time_t problems (when viewing message headers with the 'h' sub-op command).

+++ sbbs3/msgdate.c
@@ -2,13 +2,13 @@
 
 /* Synchronet RFC822 message date/time string conversion routines */
 
-/* $Id: msgdate.c,v 1.4 2011/10/16 12:29:59 rswindell Exp $ */
+/* $Id: msgdate.c,v 1.3 2007/08/13 22:12:19 deuce Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
  * @format.use-tabs true	(see http://www.synchro.net/ptsc_hdr.html)		*
  *																			*
- * Copyright 2011 Rob Swindell - http://www.synchro.net/copyright.html		*
+ * Copyright 2006 Rob Swindell - http://www.synchro.net/copyright.html		*
  *																			*
  * This program is free software; you can redistribute it and/or			*
  * modify it under the terms of the GNU General Public License				*
@@ -170,7 +170,7 @@ when_t DLLCALL rfc822date(char* date)
 	}
 
 	tm.tm_isdst=-1;	/* Don't adjust for daylight-savings-time */
-	when.time=(uint32_t)mktime(&tm);
+	when.time=mktime(&tm);
 
 	return(when);
 }
